changed game activity life cycle to reconnect automatically
Verious quirks of the quests window management means that it very frequently stops the active activity. This is very annoying as the use has to make multiple inputs to reconnect. This change introduces history to the .game activity and allows it to reconnect again instead of always finishing on stop

next step would be to try to maintain the connection completely (maybe on an option)