quest window size launcher window and mouse sensitivity
added a launcher activity that relaunches the application in a new activity with a fixed layout size. This is hack to get around the quest not allowing applications to resize to specific sizes. This also bypasses the max window size (so long as the player doesnt resize it)

Also added mouse sensitivity preference